02 - Face Recognition Part 1/deployment/face_alignment/requirements.txt
dlib==19.21.0
numpy==1.18.5
Pillow==7.2.0
opencv-python-headless
requests_toolbelt